requirements.txt
Flask==2.3.3
gunicorn==22.0.0
numpy==2.0.1
Jinja2==3.1.4
python-dotenv==1.0.1
anthropic==0.33.1
requests==2.26.0
importlib-metadata==8.0.0